Increase audio level with a factor 2
Sound will be twice as loud now.

This may lead to a brownout if the battery isn't connected.

+++ Core/Inc/porting/common.h
@@ -32,4 +32,4 @@ extern int16_t pendingSamples;
 extern int16_t audiobuffer_emulator[AUDIO_BUFFER_LENGTH] __attribute__((section (".audio")));
 extern int16_t audiobuffer_dma[AUDIO_BUFFER_LENGTH * 2] __attribute__((section (".audio")));
 
-extern const int16_t volume_tbl[ODROID_AUDIO_VOLUME_MAX + 1];
+extern const uint8_t volume_tbl[ODROID_AUDIO_VOLUME_MAX + 1];
